+++ src/Dingbat/Action/Task/Update.php
@@ -4,8 +4,9 @@
 namespace Dingbat\Action\Task;
 
 use Dingbat\Action;
-use Dingbat\Model\Card;
 use Dingbat\Model\Task;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Respect\Validation\Exceptions\NestedValidationException;
 use Slim\Http\Request;
 use Slim\Http\Response;
 
@@ -14,61 +15,34 @@ class Update implements Action
 
     public function __invoke(Request $request, Response $response, array $args)
     {
-        $id = $args['id'];
-
-        /* @var Task $task */
-        $task = null;
         try {
-            $task = Task::query()->findOrFail($id);
-        } catch (\Exception $e) {
-            return $response
-                ->withStatus(404)
-                ->withJson(['message' => sprintf('task with `id` `%d` does not exist', $id)]);
-        }
+            // load and fill model
+            /* @var Task $model */
+            $model           = Task::query()->findOrFail($args['id']);
+            $model->name     = $request->getParsedBodyParam('name');
+            $model->marked   = (bool) $request->getParsedBodyParam('marked');
+            $model->priority = $request->getParsedBodyParam('priority', Task::PRIORITY_NORMAL);
+            $model->cardId   = (int) $request->getParsedBodyParam('cardId');
 
-        // check if cardId is set
-        if ($request->getParsedBodyParam('cardId', false) === false)
-        {
-            return $response
-                ->withStatus(400)
-                ->withJson(['message' => 'param `cardId` is required']);
-        }
+            // validation
+            Task::validators()['name']->assert($model->name);
+            Task::validators()['marked']->assert($model->marked);
+            Task::validators()['priority']->assert($model->priority);
+            Task::validators()['cardId']->assert($model->cardId);
 
-        // check if cardId is exist
-        $cardId = $request->getParsedBodyParam('cardId');
-        if (Card::query()->find($cardId) === null)
-        {
-            return $response
-                ->withStatus(400)
-                ->withJson(['message' => sprintf('card with id `%d` does not exist', $cardId)]);
-        }
-
-        // check if `name` is set
-        if ($request->getParsedBodyParam('name', false) === false)
-        {
-            return $response
-                ->withStatus(400)
-                ->withJson(['message' => 'param `name` is required']);
-        }
-
-        // check if `priority` value
-        if (!in_array($request->getParsedBodyParam('priority', 'normal'), ['normal', 'high', 'low']))
-        {
-            return $response
-                ->withStatus(400)
-                ->withJson(['message' => 'param `priority` must be `normal`, `high` or `low`']);
-        }
-
-        // save task
-        try
-        {
-            $task->name     = $request->getParsedBodyParam('name');
-            $task->marked   = $request->getParsedBodyParam('marked');
-            $task->priority = $request->getParsedBodyParam('priority', Task::PRIORITY_NORMAL);
-            $task->cardId   = $cardId;
-            $task->saveOrFail();
+            // save
+            $model->saveOrFail();
 
             return $response->withStatus(204);
+        } catch (ModelNotFoundException $e) {
+            return $response
+                ->withStatus(404)
+                ->withJson(['message' => sprintf('task with `id` `%d` does not exist', $args['id'])]);
+        } catch (NestedValidationException $e) {
+            // validation error
+            return $response
+                ->withStatus(400)
+                ->withJson(['message' => $e->getFullMessage()]);
         } catch (\Exception $e) {
             return $response
                 ->withStatus(500)